split pigeon_lib to internal/external
